refactor: create nav item files
