+++ src/components/ui/Nav/index.ts
@@ -1,0 +1,2 @@
+export * from './NavItem';
+export * from './NavItemMobile';
